Proactively renaming RageDisplay::PixelFormat to RageDisplay::RagePixelFormat to resolve naming conflict with newer versionf of FFMpeg. Resolving problem with RageSoundResampler that was causing build failures on 64-bit machines.

+++ src/RageSoundResampler.h
@@ -38,7 +38,7 @@ public:
 	/* Read converted data.  Returns the number of bytes filled.
 	 * If eof() has been called and the output is completely
 	 * flushed, returns -1. */
-	int read(void *data, unsigned bytes);
+	int read(void *data, size_t bytes);
 };
 
 #endif
